--- accounts/.eslintrc.js
@@ -1,51 +1,0 @@
-module.exports = {
-  env: {
-    es2021: true,
-    node: true,
-  },
-  extends: 'standard-with-typescript',
-  overrides: [
-  ],
-  parserOptions: {
-    ecmaVersion: 'latest',
-    sourceType: 'module',
-    project: 'tsconfig.json',
-  },
-  rules: {
-    'space-before-function-paren': [
-      'error', {
-        anonymous: 'always',
-        named: 'never',
-        asyncArrow: 'always',
-      },
-    ],
-    '@typescript-eslint/space-before-function-paren': [
-      'error', {
-        anonymous: 'always',
-        named: 'never',
-        asyncArrow: 'always',
-      },
-    ],
-    semi: [2, 'always'],
-    '@typescript-eslint/semi': [2, 'always'],
-    'comma-dangle': ['error', {
-      arrays: 'always-multiline',
-      objects: 'always-multiline',
-      imports: 'always-multiline',
-      exports: 'always-multiline',
-      functions: 'never',
-    }],
-    '@typescript-eslint/comma-dangle': ['error', {
-      arrays: 'always-multiline',
-      objects: 'always-multiline',
-      imports: 'always-multiline',
-      exports: 'always-multiline',
-      functions: 'never',
-    }],
-    '@typescript-eslint/member-delimiter-style': ['error', {
-      multiline: { delimiter: 'semi', requireLast: true },
-      singleline: { delimiter: 'semi', requireLast: false },
-      multilineDetection: 'brackets',
-    }],
-  },
-};
