+++ accounts/.eslintrc.yml
@@ -1,0 +1,14 @@
+env:
+  commonjs: true
+  es2021: true
+  node: true
+extends: airbnb-base
+overrides: []
+parserOptions:
+  ecmaVersion: latest
+rules:
+  max-classes-per-file: off
+  no-underscore-dangle:
+    - error
+    - allow:
+      - _id
