+++ accounts/package.json
@@ -2,12 +2,11 @@
   "name": "accounts",
   "version": "1.0.0",
   "description": "Module to handle accounts, users and authentication/authorization",
-  "main": "./dist/src/app.js",
+  "main": "./src/app.js",
   "scripts": {
-    "start": "node ./dist/app.js",
-    "build": "rm -r ./dist/ && tsc --build",
-    "test": "NODE_ENV=test npm run build && jest",
-    "dev": "NODE_ENV=dev nodemon ./src/app.ts"
+    "start": "node ./src/app.js",
+    "test": "NODE_ENV=test jest",
+    "dev": "NODE_ENV=dev nodemon ./src/app.js"
   },
   "repository": {
     "type": "git",
@@ -29,20 +28,9 @@
     "winston": "^3.8.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.7",
-    "@babel/preset-env": "^7.20.2",
-    "@types/bcryptjs": "^2.4.2",
-    "@types/cors": "^2.8.13",
-    "@types/express": "^4.17.15",
-    "@types/jest": "^29.2.4",
-    "@types/validator": "^13.7.10",
-    "babel-jest": "^29.3.1",
-    "eslint": "^8.30.0",
-    "eslint-config-standard-with-typescript": "^24.0.0",
+    "eslint": "^8.31.0",
+    "eslint-config-airbnb-base": "^14.2.1",
     "jest": "^29.3.1",
-    "nodemon": "^2.0.20",
-    "ts-jest": "^29.0.3",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.9.4"
+    "nodemon": "^2.0.20"
   }
 }
