+++ accounts/src/models/account.js
@@ -1,0 +1,4 @@
+const { model } = require('mongoose');
+const { AccountSchema } = require('./schemas');
+
+module.exports = model('account', AccountSchema);
