+++ accounts/src/models/index.js
@@ -1,0 +1,9 @@
+const account = require('./account');
+const schemas = require('./schemas');
+const user = require('./user');
+
+module.exports = {
+  account,
+  schemas,
+  user,
+};
