--- accounts/src/models/index.ts
@@ -1,4 +1,0 @@
-export * from './types';
-export * from './account';
-export * from './user';
-export * from './account';
