+++ accounts/src/models/user.js
@@ -1,0 +1,4 @@
+const { model } = require('mongoose');
+const { UserSchema } = require('./schemas');
+
+module.exports = model('user', UserSchema);
