--- accounts/src/models/user.ts
@@ -1,5 +1,0 @@
-import { model } from 'mongoose';
-import { UserSchema } from './schemas';
-import { IUser } from './types';
-
-export const User = model<IUser>('user', UserSchema);
