+++ accounts/src/util/status-code.js
@@ -1,0 +1,10 @@
+module.exports = {
+  BAD_REQUEST: 400,
+  UNAUTHENTICATED: 401,
+  UNAUTHORIZED: 403,
+  NOT_FOUND: 404,
+  INTERNAL: 500,
+  SUCCESS: 200,
+  CREATED: 201,
+  NO_CONTENT: 204,
+};
